Separate data generation and csv writing
